+++ src/Render/DX12/StdDXRPipeline.cpp
@@ -1480,7 +1480,7 @@ void StdDXRPipeline::Impl::CameraRender(const RenderContext& ctx, const CameraDa
 	size_t width = rtb->GetDesc().Width;
 	size_t height = rtb->GetDesc().Height;
 	CD3DX12_VIEWPORT viewport(0.f, 0.f, width, height);
-	D3D12_RECT scissorRect(0.f, 0.f, width, height);
+	D3D12_RECT scissorRect(0.f, 0.f, static_cast<LONG>(width), static_cast<LONG>(height));
 
 	UpdateCameraResource(cameraData, width, height);
